rpi/requirements.txt
tflite-runtime
scipy
pyserial
numpy
